# openapi.yaml
openapi: 3.1.0
info:
  title: Badminton Court Booking API
  # The version of this DOCUMENT. Not the same as the v1 in the URL.
  version: 1.2.1
  description: |
    Books badminton courts at a university sports hall.

    It is written for three kinds of caller: **students**, who browse the
    courts and book, read and cancel their own slots; **administrators**,
    who manage the courts themselves and can act on any booking; and a
    **scheduled job** that runs nightly with no user present to tidy up
    expired bookings.

    ## Conventions

    All timestamps are RFC 3339 with an explicit UTC offset, for example
    `2026-03-04T12:10:04+07:00`. A booking covers a half-open interval: a
    booking that ends at 19:00 does not overlap one that starts at 19:00.

    All identifiers are opaque and server-generated, prefixed by the kind of
    thing they name — `crt_` for a court, `bkg_` for a booking. They are
    stable forever. Clients MUST NOT parse them or infer ordering from them.

    Clients MUST ignore unknown response fields. Clients MUST treat an
    unrecognised enumeration value as the nearest non-terminal state and
    MUST NOT crash on it. These two sentences are what make every additive
    change to this interface legal.

    This service carries no monetary values. Courts are booked, not bought.

    ## Access control

    Every operation except `GET /health` requires an OAuth 2 access token,
...

...
       declares — a missing scope is `403`;
...

...

...

    ## Idempotency
...
servers:
...
security:
  - oauth2: [courts:read]
paths:
...
  /courts:
    get:
      operationId: listCourts
      summary: List badminton courts
      tags: [Courts]
      description: |
        Returns every court, in both statuses. Retired courts are included
        on purpose, so a client can show them greyed out with a reason
        rather than having them vanish between one page load and the next.
        Filter with `status=active` when only bookable courts are wanted.

        `isAvailable` and `status` answer different questions and a client
        needs both. `status: retired` is permanent — the court is gone from
        service. `isAvailable: false` on an `active` court is temporary,
        for example while it is being resurfaced. Only a court that is
        `active` **and** `isAvailable` can be booked; either one being
        false makes `POST /v1/bookings` answer `409`.

        The list is not ordered by any field a client should rely on
        beyond the pagination cursor. Removing the `status` filter returns
        more courts, never a different kind of thing.
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
        - name: status
          in: query
          required: false
          description: Filter to things in one status.
          schema:
            type: string
            enum: [active, retired]
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          required: false
          description: Opaque cursor from a previous page's nextCursor.
          schema: { type: string }
      responses:
        '304':
          $ref: '#/components/responses/NotModified'
        '200':
          description: A page of badminton courts.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Court' }
                  nextCursor:
                    type: string
                    description: Absent on the last page.
                    example: eyJvIjoyMH0
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /courts/{courtId}:
    parameters:
      - name: courtId
        in: path
        required: true
        description: Opaque, server-generated. Never a database integer.
        schema:
          type: string
          pattern: '^crt_[A-Za-z0-9]{3,}$'
        example: crt_51Fa93cD
    get:
      operationId: getCourt
      summary: Fetch one badminton court
      tags: [Courts]
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '304':
          $ref: '#/components/responses/NotModified'
        '200':
          description: The badminton court.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Court' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # -------------------------------------------------------------------
  # Rule 5 — a state transition is a sub-resource, and its URI is a
  # NOUN. Not /markRetired, not PATCH {"status":"retired"}.
  # -------------------------------------------------------------------
...
components:
...
  securitySchemes:
    # One scheme carrying both flows, rather than two schemes. The
    # scheduled job and the two web clients present the same kind of token
    # to the same API; what differs is only how each one obtained it. Split
    # across two schemes, the client-credentials half was referenced by no
    # operation, which the linter reported as an unused component — an
    # accurate description of a flow nothing could actually use.
    oauth2:
      type: oauth2
      description: |
        Access tokens are issued by the Authorisation Server. Public
        clients use Authorization Code + PKCE with no client secret. The
        scheduled cleanup job has no user present, so it authenticates
        itself with Client Credentials and a secret held in a secret
        manager — never in this repository.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.example.com/realms/badminton/protocol/openid-connect/auth
          tokenUrl: https://auth.example.com/realms/badminton/protocol/openid-connect/token
          refreshUrl: https://auth.example.com/realms/badminton/protocol/openid-connect/token
          scopes:
            courts:read: Browse badminton courts and availability
            courts:write: Create, update, or retire badminton courts
            bookings:read: Read bookings visible to the principal
            bookings:write: Create and cancel own bookings
            bookings:fulfil: Confirm, reject, or manage all bookings
        clientCredentials:
          tokenUrl: https://auth.example.com/realms/badminton/protocol/openid-connect/token
          scopes:
            bookings:read: Read bookings for the scheduled cleanup job
            bookings:fulfil: Manage all bookings for the scheduled cleanup job
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: |
        The `ETag` of the version the client already holds. When it still
        matches, the service answers `304 Not Modified` with no body —
        which is a successful read, not a failure.

        A client polling a collection sends this on every cycle. Without
        it the service has no way to tell that the caller already has
        the current version, and every poll transfers the whole list.
      schema:
        type: string
        example: '"v8m6WCTfmEMOvl8xpz5syHmURr7"'
...
  schemas:
...
    Court:
      type: object
      required: [id, name, location, courtType, isAvailable, status]
      properties:
        id:
          type: string
          pattern: '^crt_[A-Za-z0-9]{3,}$'
          description: Opaque, server-generated, stable forever.
          example: crt_51Fa93cD
        name:
          type: string
          example: Court 1
        location:
          type: string
          example: UGM Sports Hall
        courtType:
          type: string
          enum: [indoor, outdoor]
          example: indoor
        isAvailable:
          type: boolean
          example: true
        status:
          type: string
          enum: [active, retired]
          description: Lifecycle status of the badminton court.
          example: active
...
    Problem:
      type: object
      required: [type, title, status]
      # RFC 9457 permits EXTENSION MEMBERS: your own fields alongside the
      # five below. They are what turns a merely correct error into an
      # interface a client can build a good experience on, and they are
      # explicitly assessed in A1 — so the schema must allow them.
      # Remove this line and your own extension members become lint
      # errors, which is exactly how a team ends up with useless errors.
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri
          description: |
            Stable URI identifying the KIND of problem. Client code
            branches on this exact string, so it must never change
            once published.
          example: https://api.example.com/problems/illegal-transition
        title:
          type: string
          description: |
            Short, human-readable, FIXED for a given type. Written for
            a developer reading a log, not for an end user.
          example: That status change is not permitted
        status:
          type: integer
          description: The HTTP status, repeated so the document stays
            meaningful when logged away from its response.
          example: 409
        detail:
          type: string
          description: |
            Specific to this occurrence. Never a stack trace, an SQL
            fragment, or an internal hostname.
          example: thg_51Fa93cD is draft; retirement is refused.
        instance:
          type: string
          description: |
            The URI of this specific occurrence. In Chapter 9 the
            correlation identifier goes here.
          example: /v1/things/thg_51Fa93cD/retirement
        invalid-params:
          type: array
          description: |
            Which fields were refused and why — the extension member
            RFC 9457 uses in its own example. Present on schema
            failures (400) and on cross-field rules (422).

            Each entry names one field and carries a reason written for
            the person filling the form in, so a client can place the
            message on that field rather than inventing wording for a
            rule only this service knows. A client that receives this
            member MUST NOT assume the list is exhaustive: fields are
            reported as they are checked.
          items:
            type: object
            required: [name, reason]
            properties:
              name:
                type: string
                description: |
                  The field the problem is about, as named in the
                  request body or query string.
                example: endTime
              reason:
                type: string
                description: Why it was refused, in domain terms.
                example: The end time must be after the start time
  headers:
    ETag:
      description: |
        Opaque version marker for this exact representation. A client
        stores it and sends it back in `If-None-Match` on the next read
        (A.7), and in `If-Match` on a write against the same entity
        (A.8).

        Derived from the representation, so it changes when, and only
        when, what the client would receive changes. It is strong: two
        responses carrying the same tag are byte-for-byte identical.
      required: true
      schema:
        type: string
        example: '"v8m6WCTfmEMOvl8xpz5syHmURr7"'
    CacheControl:
      description: |
        `private, no-cache`. These representations are shaped by who
        asked — a student's booking list is theirs alone — so a shared
        cache must not reuse one for somebody else. `no-cache` means
        "revalidate before reusing", which is the conversation `ETag`
        and `If-None-Match` are for.
      schema:
        type: string
        example: private, no-cache
  responses:
    NotModified:
      description: |
        The representation has not changed since the version named by
        `If-None-Match`. No body is sent.

        This is a successful read. A client treats it as confirmation
        that what it already holds is current — clearing any stale
        marker — and not as a failure.
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
    BadRequest:
      description: The request was malformed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.example.com/problems/malformed-request
            title: The request could not be parsed
            status: 400
            detail: One or more fields do not match the documented schema
            invalid-params:
              - name: startTime
                reason: Give a start date and time
    Unauthorized:
      description: The token is absent, unverifiable, or expired.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          example: 'Bearer error="invalid_token"'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: The token is valid but does not carry the scope this operation requires.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          example: 'Bearer error="insufficient_scope", scope="bookings:write"'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: |
        The object does not exist, or it exists and is not accessible to this caller.
        Both conditions are answered identically so identifiers cannot be enumerated.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.example.com/problems/not-found
            title: Resource not found
            status: 404
...